--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,12 +16,14 @@ lazy_static = "1.4.0"
 openssl = "0.10.41"
 postgres = "0.19.2"
 postgres-openssl = "0.5.0"
-rand = "0.8"
+rand = { version = "0.8", features = ["small_rng"] }
 regex = "1"
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 structopt = "0.3"
 uuid = { version = "0.8", features = [ "v4"] }
+fnv = "1"
+mimalloc = "0.1.29"
 
 [dev-dependencies]
 pretty_assertions = "1.2.1"