--- a/src/parsers/strategy_file.rs
+++ b/src/parsers/strategy_file.rs
@@ -151,9 +151,12 @@ pub fn to_csv(strategy_file: &str, csv_output_file: &str) -> std::io::Result<()>
 
 fn read_file(file_name: &str) -> Result<Vec<StrategyInFile>, std::io::Error> {
     let result = fs::read_to_string(file_name).map(|file_contents| {
-        let p: Vec<StrategyInFile> = serde_json::from_str(&file_contents)
-            .unwrap_or_else(|_| panic!("Invalid json found in strategy file at '{}'", file_name));
-        p
+        serde_json::from_str::<Vec<StrategyInFile>>(&file_contents).unwrap_or_else(|e| {
+            panic!(
+                "Invalid json found in strategy file at '{}': {:#}",
+                file_name, e
+            )
+        })
     });
 
     match result {