This builds off of https://github.com/Multiverse-io/anonymiser/pull/15, so that should go in first to make the changes here more clear:

Changed the default RNG source to SmallRng - a good comparison is here https://rust-random.github.io/book/guide-rngs.html, there may be even faster ones but the key point here is that we don't need a cryptographically secure implementation so its OK to use a higher performance once

Originally I changed some frequently used hashmaps to use Fnv as the hashing algorithm - It doesn't make a huge difference in this case because the benefits are only seen with very small or large keys (see graphs here https://cglab.ca/~abeinges/blah/hash-rs/), but as an example the average table name length (used as a hashmap key) is 28 bytes so not really any benefit of changing it

The biggest performance win was changing the default allocator to https://crates.io/crates/mimalloc 

Before:
```
rohan@Rohans-MacBook-Pro anonymiser % time cargo run --release anonymise
    Finished release [optimized + debuginfo] target(s) in 0.07s
     Running `target/release/anonymiser anonymise`
cargo run --release anonymise  33.45s user 1.62s system 97% cpu 35.921 total
```

After:
```
rohan@Rohans-MacBook-Pro anonymiser % time cargo run --release anonymise
    Finished release [optimized + debuginfo] target(s) in 0.05s
     Running `target/release/anonymiser anonymise`
cargo run --release anonymise  15.03s user 1.39s system 95% cpu 17.145 total
```
35 -> 17 seconds running on an already anonymised copy of the production DB

There might be more to be done in terms of perf but parallelization is probably the next big thing to do, `rayon` could be a good thing to try if we can find some existing iterators to parallelize. 

I've also used the `anyhow` crate before and it makes error handling a lot easier in some cases as well so that might be something else I can look into