@danturn 2022-08-08
> Changed the default RNG source to SmallRng - a good comparison is here https://rust-random.github.io/book/guide-rngs.html, there may be even faster ones but the key point here is that we don't need a cryptographically secure implementation so its OK to use a higher performance once

is this true? i guess what we're trying to do here is make sure that if we were to lose a laptop or the data on it, that the data is unrecoverable to the prod state, is a non crypto secure version enough? it possibly is? but im not sure how much dfference this change makes?