@danturn 2022-08-08
> 35 -> 17 seconds running on an already anonymised copy of the production DB

amazing, wow!!! 

what is the downside to mimalloc? is there a case where its worse? or do ppl generally just use that instead of the default?