@hoyon 2022-08-08
> is this true? i guess what we're trying to do here is make sure that if we were to lose a laptop or the data on it, that the data is unrecoverable to the prod state, is a non crypto secure version enough? it possibly is? but im not sure how much dfference this change makes?

The difference is that the randomly generated strings could be more predictable than a cryptographically secure rng but has no relation to the original data.