@rohan-b99 2022-08-08
> 
> what is the downside to mimalloc? is there a case where its worse? or do ppl generally just use that instead of the default?

I'm not aware of any major downsides other than that its not super popular, there's also jemalloc https://crates.io/crates/jemallocator which is a bit more popular and might give a similar perf increase but I think doesn't support windows
I used mimalloc on a previous project than ran on x86_64 mac/windows/linux and ppc64le linux and didn't have any issues with it - there's some performance numbers here https://github.com/microsoft/mimalloc#Performance

As @hoyon said the rng method only affects the quality of the random data, so yeah I don't think anything would be recoverable since it's being replaced entirely